fixup! feat(ivy): Change static priority resolution to be same level as directive it belongs to
